requirements.txt
django
djangorestframework
django-filter
requests
psycopg2
pillow